services/agentpy/requirements.txt
fastapi
uvicorn[standard]
httpx>=0.25.0
pydantic>=2.10,<3
openai-agents==0.0.14